+++ tests/test_session.py
@@ -8,8 +8,9 @@ import spotify.manager.session
 # monkeypatch for testing
 spotify.manager.session.spotify = spotify._mockspotify
 
+from spotify import Settings
 from spotify.manager import SpotifySessionManager
-from spotify._mockspotify import mock_track, mock_album
+from spotify._mockspotify import mock_track, mock_album, Session
 
 
 class BaseMockClient(SpotifySessionManager):
@@ -21,7 +22,6 @@ class BaseMockClient(SpotifySessionManager):
 
     def __init__(self):
         SpotifySessionManager.__init__(self, "username_good", "password_good")
-        self.awoken = threading.Event() # used to block until awoken
 
 class BaseMockUnicodeClient(SpotifySessionManager):
 
@@ -32,16 +32,20 @@ class BaseMockUnicodeClient(SpotifySessionManager):
 
     def __init__(self):
         SpotifySessionManager.__init__(self, "username_good", "password_good")
-        self.awoken = threading.Event() # used to block until awoken
 
 class TestSession(unittest.TestCase):
+
+    def test_create(self):
+        c = BaseMockClient()
+        s = Settings()
+        s.application_key = "appkey_good"
+        session = Session.create(c, s)
 
     def test_initialisation(self):
         class MockClient(BaseMockClient):
             def logged_in(self, session, error):
                 username = session.username()
                 self.found_username = username
-                session.logout()
                 self.disconnect()
 
         c = MockClient()
@@ -53,7 +57,6 @@ class TestSession(unittest.TestCase):
             def logged_in(self, session, error):
                 username = session.username()
                 self.found_username = username
-                session.logout()
                 self.disconnect()
 
         c = MockClient()
@@ -67,7 +70,6 @@ class TestSession(unittest.TestCase):
                 session.load(track)
                 session.seek(40000)
                 session.play(True)
-                session.logout()
                 self.disconnect()
 
             def music_delivery(self, sess, mformat, frames):